Extract legacy view into views module
This probably isn't needed any more, but for the sake of changing one
thing at a time, we'll keep it around for now.
