+++ h/features/__init__.py
@@ -5,7 +5,6 @@ import logging
 import os
 
 from pyramid import events
-from pyramid.view import view_config
 import sqlalchemy as sa
 import transaction
 
@@ -213,20 +212,10 @@ def remove_old_flags_on_boot(event):
     transaction.commit()
 
 
-# Deprecated dedicated endpoint for feature flag data,
-# kept for compatibility with older clients (<= 0.8.6).
-# Newer clients get feature flag data as part of the session data
-# from the /app endpoint.
-@view_config(route_name='features_status',
-             request_method='GET',
-             accept='application/json',
-             renderer='json',
-             http_cache=(0, {'no_store': False}))
-def features_status(request):
-    return request.feature.all()
-
-
 def includeme(config):
+    config.include('h.features.views')
+
     config.add_request_method(Client, name='feature', reify=True)
     config.add_route('features_status', '/app/features')
+
     config.scan(__name__)
